Refactor CombatZoneState and MeteorSwarmState: rename variables for clarity; update protection logic and method signatures

+++ src/main/java/Model/State/CombatZoneState.java
@@ -18,13 +18,14 @@ public class CombatZoneState extends State {
     private PlayerData minPlayerEngines;
     private PlayerData minPlayerCannons;
     private Integer dice;
-    private Boolean useEnergy;
+    private Boolean protect;
     private Integer batteryID;
     List<List<Pair<Integer, Integer>>> fragments;
+    private Map<Integer, Integer> cabinsID;
     private Integer fragmentChoice;
     private Pair<Component, Integer> protectionResult;
     private int hitIndex;
-    private int fightIndex;
+    private int fightState;
 
     /**
      * Constructor whit players and card
@@ -36,13 +37,13 @@ public class CombatZoneState extends State {
         this.card = card;
         this.stats = new ArrayList<>(3);
         this.subState = CombatZoneStatsType.CREW;
-        this.useEnergy = null;
+        this.protect = null;
         this.fragmentChoice = null;
         this.dice = null;
         this.minPlayerCannons = null;
         this.batteryID = null;
         this.hitIndex = 0;
-        this.fightIndex = 0;
+        this.fightState = 0;
     }
 
     /**
@@ -68,7 +69,7 @@ public class CombatZoneState extends State {
      */
     private void transitionHit() {
         hitIndex++;
-        fightIndex = 0;
+        fightState = 0;
     }
 
     /**
@@ -116,17 +117,17 @@ public class CombatZoneState extends State {
      * Execute the subState engines
      * @param player player to execute
      */
-    private void executeSubStateEngines(PlayerData player) {
+    private void executeSubStateEngines(PlayerData player) throws IllegalArgumentException {
         int crewLost = card.getLost();
-        int statsIndex = CombatZoneStatsType.ENGINES.getIndex(card.getCardLevel());
 
-        stats.get(statsIndex).entrySet().stream().min(this::comparePlayers).ifPresent(entry -> {
-            SpaceShip spaceShip = entry.getKey().getSpaceShip();
-            spaceShip.addCrewMember(-crewLost);
-            if (spaceShip.getCrewNumber() <= crewLost) {
-                player.setGaveUp(true);
+        SpaceShip spaceShip = minPlayerEngines.getSpaceShip();
+        if (spaceShip.getCrewNumber() <= crewLost) {
+            player.setGaveUp(true);
+        } else {
+            for (Map.Entry<Integer, Integer> cabinID : cabinsID.entrySet()) {
+                spaceShip.removeCrewMember(cabinID.getKey(), cabinID.getValue());
             }
-        });
+        }
 
         transition(CombatZoneStatsType.CANNONS);
     }
@@ -139,53 +140,44 @@ public class CombatZoneState extends State {
         Component component = protectionResult.getValue0();
         int protectionType = protectionResult.getValue1();
 
-        switch (protectionType) {
-            case -1:
+        if (protectionType == 0 || protectionType == -1) {
+            if (protect) {
+                spaceShip.useEnergy(batteryID);
+                transitionHit();
+            } else {
                 fragments = destroyComponent(spaceShip, component);
                 if (fragments.size() > 1) {
-                    fightIndex++;
+                    fightState++;
                 } else {
                     transitionHit();
                 }
-                break;
-            case 0:
-                if (useEnergy) {
-                    spaceShip.useEnergy(batteryID);
-                    transitionHit();
-                } else {
-                    fragments = destroyComponent(spaceShip, component);
-                    if (fragments.size() > 1) {
-                        fightIndex++;
-                    } else {
-                        transitionHit();
-                    }
-                }
-                break;
-            case 1:
-                break;
+            }
         }
     }
 
     /**
      * Execute the subState cannons
      * @param player player to execute
-     * @throws IllegalStateException if dice not set, useEnergy not set, fragmentChoice not set
+     * @throws IllegalStateException if dice not set, protect not set, fragmentChoice not set
      */
-    private void executeSubStateCannons(PlayerData player) throws IllegalStateException {
+    private void executeSubStateFight(PlayerData player) throws IndexOutOfBoundsException, IllegalStateException {
         SpaceShip spaceShip = player.getSpaceShip();
-        Hit hit = card.getFires().get(hitIndex);
 
-        switch (fightIndex) {
+        switch (fightState) {
             case 0:
+                if (hitIndex >= card.getFires().size()) {
+                    throw new IndexOutOfBoundsException("Hit index out of bounds");
+                }
                 if (dice == null) {
                     throw new IllegalStateException("Dice not set");
                 }
+                Hit hit = card.getFires().get(hitIndex);
                 protectionResult = spaceShip.canProtect(dice, hit);
-                fightIndex++;
+                fightState++;
                 break;
             case 1:
-                if (useEnergy == null) {
-                    throw new IllegalStateException("UseEnergy not set");
+                if (protect == null) {
+                    throw new IllegalStateException("Protect not set");
                 }
                 executeSubStateProtection(spaceShip);
                 break;
@@ -207,7 +199,7 @@ public class CombatZoneState extends State {
      * @throws IllegalStateException if not in the right state in order to do the action
      */
     public void setFragmentChoice(int fragmentChoice) throws IllegalStateException {
-        if (subState != CombatZoneStatsType.CANNONS || fightIndex != 2) {
+        if (subState != CombatZoneStatsType.CANNONS || fightState != 2) {
             throw new IllegalStateException("Fragment choice not allowed in this state");
         }
         this.fragmentChoice = fragmentChoice;
@@ -215,18 +207,18 @@ public class CombatZoneState extends State {
 
     /**
      * Set the use energy
-     * @param useEnergy_ use energy
+     * @param protect_ use energy
      * @param batteryID_ battery ID
      * @throws IllegalStateException if not in the right state in order to do the action
-     * @throws IllegalArgumentException if batteryID_ is null and useEnergy_ is true
+     * @throws IllegalArgumentException if batteryID_ is null and protect_ is true
      */
-    public void setUseEnergy(boolean useEnergy_, Integer batteryID_) throws IllegalStateException, IllegalArgumentException {
-        if (subState != CombatZoneStatsType.CANNONS || fightIndex != 1) {
+    public void setProtect(boolean protect_, Integer batteryID_) throws IllegalStateException, IllegalArgumentException {
+        if (subState != CombatZoneStatsType.CANNONS || fightState != 1) {
             throw new IllegalStateException("Battery ID not allowed in this state");
         }
-        this.useEnergy = useEnergy_;
-        if (useEnergy_ && batteryID_ == null) {
-            throw new IllegalArgumentException("If you set useEnergy to true, you have to set the batteryID");
+        this.protect = protect_;
+        if (protect_ && batteryID_ == null) {
+            throw new IllegalArgumentException("If you set protect to true, you have to set the batteryID");
         }
         this.batteryID = batteryID_;
     }
@@ -236,10 +228,31 @@ public class CombatZoneState extends State {
      * @param dice dice
      */
     public void setDice(int dice) {
-        if (subState != CombatZoneStatsType.CANNONS || fightIndex != 0) {
+        if (subState != CombatZoneStatsType.CANNONS || fightState != 0) {
             throw new IllegalStateException("Dice not allowed in this state");
         }
         this.dice = dice;
+    }
+
+    /**
+     * Set the cabins ID
+     * @param cabinsID Map of cabins ID and number of crew removed for cabins
+     * @throws IllegalStateException if not in the right state in order to do the action
+     */
+    public void setCabinsID(Map<Integer, Integer> cabinsID) throws IllegalStateException {
+        if (subState != CombatZoneStatsType.ENGINES) {
+            throw new IllegalStateException("setCabinsID not allowed in this state");
+        }
+
+        int crewRemoved = 0;
+        for (Map.Entry<Integer, Integer> cabinID : cabinsID.entrySet()) {
+            crewRemoved += cabinID.getValue();
+        }
+
+        if (crewRemoved != card.getLost()) {
+            throw new IllegalStateException("The crew removed is not equal to the crew lost");
+        }
+        this.cabinsID = cabinsID;
     }
 
     /**
@@ -291,12 +304,12 @@ public class CombatZoneState extends State {
             this.addDefaultStats(CombatZoneStatsType.CREW, value0, (float) value0.getSpaceShip().getCrewNumber());
 
             this.addDefaultStats(CombatZoneStatsType.ENGINES, value0, (float) value0.getSpaceShip().getSingleEnginesStrength());
-            if (value0.getSpaceShip().getBrownAlien()) {
+            if (value0.getSpaceShip().hasBrownAlien()) {
                 this.addDefaultStats(CombatZoneStatsType.ENGINES, value0, SpaceShip.getAlienStrength());
             }
 
             this.addDefaultStats(CombatZoneStatsType.CANNONS, value0, value0.getSpaceShip().getSingleCannonsStrength());
-            if (value0.getSpaceShip().getPurpleAlien()) {
+            if (value0.getSpaceShip().hasPurpleAlien()) {
                 this.addDefaultStats(CombatZoneStatsType.CANNONS, value0, SpaceShip.getAlienStrength());
             }
         }
@@ -317,13 +330,13 @@ public class CombatZoneState extends State {
                 if (minPlayerEngines == null) {
                     throw new IllegalStateException("Not all player have set their engines");
                 }
-                executeSubStateEngines(player);
+                executeSubStateEngines(minPlayerEngines);
                 break;
             case CombatZoneStatsType.CANNONS:
                 if (minPlayerCannons == null) {
                     throw new IllegalStateException("Not all player have set their cannons");
                 }
-                executeSubStateCannons(player);
+                executeSubStateFight(minPlayerCannons);
                 break;
         }
     }
